game screen separate class
