+++ src/main/java/tetris/View.java
@@ -77,8 +77,8 @@ public class View extends Application {
         controller = new Controller(this);
         this.primaryStage = primaryStage;
         gotoLauncher();
-//        primaryStage.setTitle("Tetris");
-//        primaryStage.show();
+        // primaryStage.setTitle("Tetris");
+        // primaryStage.show();
     }
 
     public void resetFocus() {
@@ -125,70 +125,8 @@ public class View extends Application {
      * fires up the game controller.
      */
     public void gotoGameScreen() {
-        scoreLabel = new ObservingLabel("0");
-        scoreLabel
-            .setStyle("-fx-background-color:red;-fx-text-fill:black;-fx-text-alignment:center;"
-                + "-fx-alignment:center;-fx-font-weight:bold;-fx-font-size:250%");
-        scoreLabel.setMinSize(100, 50);
-        Pane leftPane = setUpLeftPaneGameScreen();
-        GridPane rightPane = setUpRightPaneGameScreen();
-        GridPane rootGameScreen = setUpRootPaneGameScreen(leftPane, rightPane);
-
-        Scene gameScreen = new Scene(rootGameScreen);
-        primaryStage.setScene(gameScreen);
-        gameScreen.getRoot().requestFocus();
-    }
-
-    private Pane setUpLeftPaneGameScreen() {
-        Canvas canvas = new Canvas(BLOCK_SIZE * BOARD_WIDTH, BLOCK_SIZE * BOARD_HEIGHT);
-        board = canvas.getGraphicsContext2D();
-        Pane leftPane = new Pane();
-        leftPane.getChildren().add(canvas);
-        leftPane.setStyle("-fx-background-color: grey");
-        return leftPane;
-    }
-
-    private GridPane setUpRootPaneGameScreen(Pane leftPane, GridPane rightPane) {
-        GridPane rootGameScreen = new GridPane();
-        rootGameScreen.setHgap(10);
-        GridPane.setConstraints(leftPane, 0, 0);
-        GridPane.setConstraints(rightPane, 1, 0);
-        rootGameScreen.getChildren().addAll(leftPane, rightPane);
-        rootGameScreen.setOnKeyPressed(event -> controller.handleKeyEvent(event));
-        return rootGameScreen;
-    }
-
-    private GridPane setUpRightPaneGameScreen() {
-        Button exitButton = new Button("exit");
-        Button restartButton = new Button("restart");
-        GridPane PreviewPane = setUpPreviewPane();
-        hookGameScreenEvents(exitButton, restartButton);
-
-        GridPane rightPane = new GridPane();
-        rightPane.setVgap(10);
-        rightPane.setAlignment(Pos.CENTER);
-        GridPane.setConstraints(exitButton, 0, 0);
-        GridPane.setConstraints(restartButton, 0, 1);
-        GridPane.setConstraints(scoreLabel, 0, 2);
-        GridPane.setConstraints(PreviewPane, 0, 4);
-        rightPane.getChildren().addAll(exitButton, restartButton, PreviewPane, scoreLabel);
-        rightPane.setStyle("-fx-background-color: grey");
-        return rightPane;
-    }
-
-    private GridPane setUpPreviewPane() {
-        Canvas canvas = new Canvas(BLOCK_SIZE * 6, BLOCK_SIZE * 5);
-        PreviewGC = canvas.getGraphicsContext2D();
-
-        GridPane PreviewPane = new GridPane();
-        PreviewPane.getChildren().addAll(canvas);
-
-        return PreviewPane;
-    }
-
-    private void hookGameScreenEvents(Button exitButton, Button restartButton) {
-        exitButton.setOnAction(event -> controller.stop());
-        restartButton.setOnAction(event -> controller.restartGame());
+        primaryStage.setScene(new Scene(new GameView()));
+        primaryStage.getScene().getRoot().requestFocus();
     }
 
     /**
